refactor: default relays on p2p
